Move version.go to commands dir.
